fix: don't assume vertices are uint32
dijkstra3d will return numpy arrays in the correct dtype.

+++ kimimaro/trace.py
@@ -161,7 +161,7 @@ def trace(
     [ PrecomputedSkeleton.from_path(path) for path in paths ]
   ).consolidate()
 
-  verts = skel.vertices.flatten().astype(np.uint32)
+  verts = skel.vertices.flatten()
   skel.radii = DBF[verts[::3], verts[1::3], verts[2::3]]
 
   return skel
